Improved GenerateResourceAndImage.ps1 PowerShell script
When the resource group specified by the user is already existent it will be forcefully removed without user confirmation. This can cause serious trouble!
I added a confirmation prompt asking the user how to continue in that case. In addition I put in a Force flag that skips this new behavior and just forcefully removes the resource group.

Names of resource groups may contains special characters, that are not valid for storage account names. I removed the most common ones, so that the script does not fail in that case! Moreover I adjusted the script, so that it follows the naming conventions for azure provided by microsoft.